# backend/swagger/api-docs/ram-api-docs.yaml
swagger: '2.0'
...
info:
  title: RAM Public API
  description: Full documentation of Relationship Access Manager (RAM) public API. All the listed API calls must be prefixed with /api/v1
  version: "1.0.0"
host:
schemes:
  - http
  - https
basePath: /api/v1
produces:
  - application/json

# ---------------------------------------------------------------------------------------------------------------------
# End Points
# ---------------------------------------------------------------------------------------------------------------------
paths:
...
  /party/identity/{idValue}:
    get:
      tags:
        - Business - Party
      operationId: Party.findByIdentityIdValue
      summary: Get the party by its identity id value
      description: Get the party by its identity id value
      parameters:
        - $ref: "#/parameters/idValue"
      responses:
        200:
          description: The updated party returned
          schema:
            $ref: "#/definitions/Party"
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'

  # Party Types .......................................................................................................
...
  /relationship/{id}:
    get:
      tags:
        - Business - Relationship
      operationId: Relationship.findByIdentifier
      summary: Get Relationship by identifier
      description: Get Relationship by identifier
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
    delete:
      tags:
        - Business - Relationship
      operationId: Relationship.deleteByIdentifier
      summary: Delete Relationship by Identifier
      description: Delete Relationship by Identifier
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationship/invitationCode/{invitationCode}:
    get:
      tags:
        - Business - Relationship
      operationId: Relationship.findPendingByInvitationCodeInDateRange
      summary: Get pending Relationship by invitation code
      description: Get pending Relationship by invitation code
      parameters:
        - $ref: "#/parameters/invitationCode"
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationship/invitationCode/{invitationCode}/accept:
    post:
      tags:
        - Business - Relationship
      operationId: Relationship.acceptByInvitationCode
      summary: Accepts a pending Relationship by invitation code
      description: Accepts a pending Relationship by invitation code
      parameters:
        - $ref: "#/parameters/invitationCode"
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationship/invitationCode/{invitationCode}/reject:
    post:
      tags:
        - Business - Relationship
      operationId: Relationship.rejectByInvitationCode
      summary: Rejects a pending Relationship by invitation code
      description: Rejects a pending Relationship by invitation code
      parameters:
        - $ref: "#/parameters/invitationCode"
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationship/invitationCode/{invitationCode}/notifyDelegate:
    post:
      tags:
        - Business - Relationship
      operationId: Relationship.notifyDelegate
      summary: Notifies the delegate of a Relationship by invitation code
      description: Notifies the delegate of a Relationship by invitation code
      parameters:
        - $ref: "#/parameters/invitationCode"
        - in: body
          required: true
          schema:
            $ref: '#/definitions/NotifyRelationshipDelegate'
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationships/identity/{idValue}:
    get:
      tags:
        - Business - Relationship
      operationId: Relationship.searchByIdentity
      summary: Search Relationships for an identity
      description: Search Relationships for an identity
      parameters:
        - $ref: "#/parameters/idValue"
        - name: filter
          in: query
          description: Filtering and Sorting
          required: false
          type: string
        - name: page
          in: query
          description: Page used for pagination
          required: true
          type: number
          format: int32
        - name: pageSize
          in: query
          description: Page Size used for pagination
          required: false
          type: number
          format: int32
      responses:
        200:
          schema:
            $ref: '#/definitions/RelationshipListResponse'
        400:
          $ref: '#/responses/400'
        500:
          $ref: '#/responses/500'
...
  /relationship:
    post:
      tags:
        - Business - Relationship
      summary: Create Relationship
      description: Create Relationship
      parameters:
        - in: body
          required: true
          schema:
            $ref: '#/definitions/CreateRelationship'
      responses:
        200:
          description: A relationship document
          schema:
              $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        500:
          $ref: '#/responses/500'
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        500:
          $ref: '#/responses/500'

  # Relationship Statuses..............................................................................................
  /relationshipStatus/{code}:
    get:
      tags:
        - Business - Relationship Status
      operationId: Relationship.findStatusByName
      summary: Get the relationship status by code
      description: Get the relationship status by code
      parameters:
        - $ref: "#/parameters/code"
      responses:
        200:
          description: The relationship status returned
          schema:
            $ref: "#/definitions/RelationshipStatus"
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
  /relationshipStatuses:
    get:
      tags:
        - Business - Relationship Status
      operationId: Relationship.listStatuses
      summary: Get the relationship statuses
      description: Get the relationship statuses
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/RelationshipStatusResourceReference'
        500:
          $ref: '#/responses/500'

  # -------------------------------------------------------------------------------------------------------------------
  # Reference End Points
  # -------------------------------------------------------------------------------------------------------------------

  # Relationship Types ................................................................................................
...
definitions:
...
  Name:
    type: object
    description: Name of a party of profile for either an individual or organisation
    properties:
      givenName:
        type: string
        description: The given name of an individual
      familyName:
        type: string
        description: The family name of an individual
      unstructuredName:
        type: string
        description: The unstructured name of an organisation
      _displayName:
        type: string
        description: The computed name of the individual or organisation

  # Shared Secret .....................................................................................................
...
  Party:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    - $ref: '#/definitions/PartyDetails'
    description: A Party is the concept that participates in Relationships.
    type: object
    required: ["roles","identities"]
  PartyDetails:
    type: object
    required: ["partyType","identities"]
    properties:
      partyType:
        type: string
      identities:
        type: array
        items:
          $ref: '#/definitions/IdentityResourceReference'
  PartyResourceReference:
    description: A reference to a Party
    type: object
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Party
      value:
          $ref: '#/definitions/Party'
...
  IdentityResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Identity
      value:
        $ref: '#/definitions/Identity'

  # Relationship ......................................................................................................
  Relationship:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    - $ref: '#/definitions/RelationshipBusinessIdentifier'
    - $ref: '#/definitions/RelationshipDetails'
    type: object
    description: |
      A Relationship is a connection between two parties.
      RAM will only record relationships if it is useful (in some way) to agencies for deciding if one party to
      confer authoriatation to another party
    required: ["relationshipTypeURI","subjectPartyRef","delegatePartyRef","startTimestamp","status"]
  RelationshipBusinessIdentifier:
    type: object
    required: ["relationshipTypeURI","subjectPartyRef","delegatePartyRef","startTimestamp"]
    properties:
      relationshipTypeURI:
        type: string
        description: URI that uniquely identifies the reference data type that gives semantic meaning to  this relationship
      subjectPartyRef:
        $ref: '#/definitions/PartyResourceReference'
      delegatePartyRef:
        $ref: '#/definitions/PartyResourceReference'
      startTimestamp:
        type: string
        description: Defines when this relationship comes into effect.  Date in ISO format.
        format: date
  RelationshipDetails:
    type: object
    required: ["status"]
    properties:
      # this has not yet been standardised to a role reference
      subjectPartyRoleURI:
        type: string
        description: URI that uniquely identifies the subject's role in this relationship
      delegatePartyRoleURI:
      # this has not yet been standardised to a role reference
        type: string
        description: URI that uniquely identifies the delegate's role in this relationship
      subjectsNickname:
          $ref: '#/definitions/Name'
      delegatesNickname:
          $ref: '#/definitions/Name'
      endTimestamp:
        type: string
        description: Date in ISO format
        format: date
      attributes:
        type: array
        description: Attributes attached to this relationship based on role type
        items:
          $ref: '#/definitions/RelationshipAttribute'
      status:
        type: string
        description: Indicates where in the lifecycle of relationships; this one is up to.
        enum: ['invalid','pending','active','deleted','cancelled']
  CreateRelationship:
    type: object
    properties:
      rawIdValue:
        type: string
        description: The identity raw id value
      partyType:
        type: string
        description: The Party Type
      givenName:
        type: string
        description: The given name
      familyName:
        type: string
        description: The family name
      unstructuredName:
        type: string
        description: The unstructured name
      sharedSecretTypeCode:
        type: string
        description: The shared secret code
      sharedSecretValue:
        type: string
        description: The schared secret value
      identityType:
        type: string
        description: The identity type
      agencyScheme:
        type: string
        description: The agency scheme
      agencyToken:
        type: string
        description: The agency token
      linkIdScheme:
        type: string
        description: The link id scheme
      linkIdConsumer:
        type: string
        description: The link id consumer
      publicIdentifierScheme:
        type: string
        description: The public identifier scheme
      profileProvider:
        type: string
        description: The profile provider
  NotifyRelationshipDelegate:
    type: object
    properties:
      email:
        type: string
        description: The email of the delegate
  RelationshipResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Relationship
      value:
        $ref: '#/definitions/Relationship'
  RelationshipListResponse:
    allOf:
    - $ref: '#/definitions/SearchResultResponse'
    type: object
    required: ["list"]
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/RelationshipResourceReference'

  # Relationship Attribute ............................................................................................
  RelationshipStatus:
    type: object
    description: Relationship status
    properties:
      code:
        type: string
        description: The code or name
      shortDecodeText:
        type: string
        description: The human friendly label
  RelationshipStatusResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this RelationshipStatus
      value:
        $ref: '#/definitions/RelationshipStatus'

  # Relationship Attribute ............................................................................................
  RelationshipAttribute:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    type: object
    description: A Relationship Attribute is a way of recording some "facts" about a Relationship
    required: ["name","value"]
    properties:
      name:
        type: string
        description: Name of an attribute definition
      attributeName:
        $ref: '#/definitions/RelationshipAttributeNameResourceReference'

  # -------------------------------------------------------------------------------------------------------------------
  # Reference Models
  # -------------------------------------------------------------------------------------------------------------------

  # Relationship Type .................................................................................................
...
  IRAMObject:
    type: object
    required: ['createdAt','updatedAt','deleteInd','resourceVersion']
    properties:
      createdAt:
        description: A timestamp when this resource was first created.
        type: string
        format: date-time
      updatedAt:
        description: A timestamp when this resource was last updated.
        type: string
        format: date-time
      deleteInd:
        type: boolean
        description: An indicated if this resource has been logically deleted.
      resourceVersion:
        type: number
        description: A schema version.
...
  SearchResultResponse:
    type: object
    description: |
      This abstract object just adds common properties to a <entity>SearchResultResponse object
    required: [totalCount, pageSize]
    properties:
      page:
        description: Current page number
        type: integer
        format: int32
      totalCount:
        description: Total number of records matched for this query
        type: integer
        format: int32
      pageSize:
        description: Page Size for the results, which the server may ignore for performance reasons
        type: integer
        format: int32
  Error:
    type: object
    required: ["alert"]
    properties:
      alert:
        $ref: '#/definitions/Alert'
  Alert:
    type: object
    required: ["alertType","messages"]
    properties:
        alertType:
            type: string
            format: can be one of the following values, success, notice or error
        messages:
            type: array
            description: Array of string presenting one or more messages
            items:
                type: string

# ---------------------------------------------------------------------------------------------------------------------
# Parameters
# ---------------------------------------------------------------------------------------------------------------------
parameters:
  code:
    name: code
    in: path
    description: The code that identifies a resource
    required: true
    type: string
...
  idValue:
    name: idValue
    in: path
    description: The unique identity composite id value
    required: true
    type: string
  invitationCode:
    name: invitationCode
    in: path
    description: The unique identity invitation code
    required: true
    type: string
...
responses:
  400:
    description: Bad Input Error
    schema:
      $ref: '#/definitions/Error'
...
  404:
    description: Not Found Error
    schema:
      $ref: '#/definitions/Error'
  500:
    description: Server Error
    schema:
      $ref: '#/definitions/Error'